changed fuel to float

+++ lander.h
@@ -19,7 +19,7 @@ class Lander {
 
   vec3 position;		// position in world coordinates (m)
   vec3 velocity;		// velocity in world coordinates (m/s)
-  int fuel = 4000;
+  float fuel = 4000;
   float orientation;		// orientation (radians CCW)
   float angularVelocity;	// angular velocity (radians/second CCW)
 
@@ -56,7 +56,7 @@ class Lander {
   void addThrust( float deltaT );
 
   vec3 centrePosition() { return position; }
-  int getFuel() { return fuel; }
+  float getFuel() { return fuel; }
   float speed() { return velocity.length(); }
   float getHorizSpeed() { return velocity[0];}
   float getVertSpeed() { return velocity[1];}
